Moved mock mail service to its own file
